✨ feat(tenon-editor, tenon-engine, tenon-materials): 支持 React 渲染

+++ packages/tenon-workbench/scripts/build.ts
@@ -1,4 +1,5 @@
 import fs from 'fs/promises';
+import syncFs from 'fs';
 import path from 'path';
 import { build } from 'vite';
 import vue from '@vitejs/plugin-vue';
@@ -9,7 +10,14 @@ import { TDesignResolver } from 'unplugin-vue-components/resolvers';
 
 
 (async () => {
-  await fs.unlink(path.resolve(__dirname, '../tsconfig.tsbuildinfo'));
+  if (syncFs.existsSync(path.resolve(__dirname, '../tsconfig.tsbuildinfo'))) {
+    await fs.unlink(path.resolve(__dirname, '../tsconfig.tsbuildinfo'));
+  }
+  if (syncFs.existsSync(path.resolve(__dirname, '../typings'))) {
+    await fs.rmdir(path.resolve(__dirname, '../typings'), {
+      recursive: true,
+    });
+  }
   await build({
     root: path.resolve(__dirname, '../'),
     base: '/',
